Cleaned up NTRU DSD implementation and added documentation
