+++ estimator/lwe_primal.py
@@ -174,10 +174,8 @@ class PrimalUSVP:
 
         """
         params = LWEParameters.normalize(params)
-
         # allow for a larger embedding lattice dimension: Bai and Galbraith
         m = params.m + params.n if params.Xs <= params.Xe else params.m
-
         if red_shape_model == "gsa":
             with local_minimum(40, max(2 * params.n, 41), precision=5) as it:
                 for beta in it:
