+++ estimator/ntru_primal.py
@@ -7,63 +7,163 @@ See :ref:`LWE Primal Attacks` for an introduction what is available.
 """
 from functools import partial
 
-from sage.all import oo, ceil, sqrt, log, RR, ZZ, binomial, cached_function
+from sage.all import oo, ceil, sqrt, log, RR, ZZ, binomial, cached_function, exp, pi, floor
+from math import lgamma
+from scipy.special import digamma
+import numpy as np
 from .reduction import delta as deltaf
 from .reduction import cost as costf
-from .util import local_minimum
+from .util import local_minimum, zeta_precomputed, zeta_prime_precomputed
 from .cost import Cost
 from .lwe_primal import PrimalUSVP, PrimalHybrid
 from .ntru_parameters import NTRUParameters
-from .overstretched_ntru import find_fatigue, combined_attack_prob
 from .errors import NotOverstretchedError
 from .simulator import normalize as simulator_normalize
 from .prob import drop as prob_drop
 from .prob import amplify as prob_amplify
 from .prob import babai as prob_babai
 from .prob import mitm_babai_probability
+from .prob import conditional_chi_squared, chisquared_table
 from .io import Logging
 from .conf import red_cost_model as red_cost_model_default
 from .conf import red_shape_model as red_shape_model_default
 from .conf import red_simulator as red_simulator_default
+from .conf import max_n_cache, gh_constant
 
 class PrimalDSD():
     """
     Estimate cost of solving (overstretched) NTRU via uSVP reduction
     """
+    @staticmethod
+    @cached_function
+    def ball_log_vol(n):
+        return RR((n/2.)) * RR(log(pi)) - RR(lgamma(n/2. + 1))
+
+    @staticmethod
+    def log_gh(d, logvol=0):
+        if d < 49:
+            return RR(gh_constant[d]) + RR(logvol)/d
+
+        return RR(1./d) * RR(logvol - PrimalDSD.ball_log_vol(d))
+
+    @staticmethod
+    def DSL_logvol_matrix(n,sigmasq):
+        total = n*(RR(log(sigmasq))+RR(log(2.))+RR(digamma(n)))/2.
+        proj_loss = np.sum([(digamma((2*n-i)/2.)-digamma(n)) for i in range(n)])/2.
+        return total+proj_loss
+
+    @staticmethod
+    def DSL_logvol_circulant(n,sigmasq):
+        lambda0 = RR((np.log(2)-np.euler_gamma+np.log(n)+np.log(sigmasq))/2.)
+        lambdai = (n-1)*(1-np.euler_gamma+np.log(n)+np.log(sigmasq))/2.
+        return lambda0+lambdai
+
+    @staticmethod
+    def DSL_logvol_circulant_fixed(n,R):
+        lambda0 = (-np.euler_gamma+np.log(R))/2.
+        lambdai = (n-1)*(1-np.euler_gamma+np.log(R)-np.log(2))/2.
+        return lambda0+lambdai
 
     @staticmethod
     @cached_function
-    def cost_simulator(
+    def DSL_logvol(n,sigmasq,ntru="circulant"):
+        if ntru=="matrix":
+            return PrimalDSD.DSL_logvol_matrix(n,sigmasq)
+        if ntru=="circulant":
+            return PrimalDSD.DSL_logvol_circulant(n,sigmasq)
+        if ntru=="fixed":
+            return PrimalDSD.DSL_logvol_circulant_fixed(n,sigmasq)
+        print("non implemented ntru type")
+
+    @staticmethod
+    @cached_function
+    def proj_logloss(d, k):
+    # log loss of length when projecting out k dimension out of d 
+        return (RR(digamma((d-k)/2.))-RR(digamma(d/2.)))/2.
+
+    @staticmethod
+    def DSLI_vols(dsl_logvol, FL_shape):
+        n = len(FL_shape)//2
+        vols = (2*n+1)*[None]
+
+        dsl_dim = n
+        vols[2*n] = dsl_logvol
+        
+        # Going to a intersection of dimension s
+        for s in range(2*n-1, n, -1):
+            # Negate cause it's a dual vector really
+            x = - FL_shape[s]
+            x += PrimalDSD.proj_logloss(s+1,n)
+            x += zeta_prime_precomputed[dsl_dim]/zeta_precomputed[dsl_dim] # primitivity
+            dsl_logvol += x
+            vols[s] = dsl_logvol
+            dsl_dim -= 1
+
+        assert(dsl_dim==1)
+        assert(s==n+1)
+        
+        return vols
+    
+    @staticmethod
+    @cached_function
+    def prob_dsd(
         beta: int,
         params: NTRUParameters,
         simulator,
         m: int = oo,
         tau=None,
         d=None,
+        dsl_logvol=None,
         red_cost_model=red_cost_model_default,
         log_level=None,
     ):
-        delta = deltaf(beta)
+        
         if d is None:
-            d = min(ceil(sqrt(params.n * log(params.q) / log(delta))), m) + 1
+            d = m
         
         xi = PrimalUSVP._xi_factor(params.Xs, params.Xe)
-
-        # TODO: this predicate isn't quite right for DSD estimate.
-        # TODO: In fact, we could replace cost simulator directly by the DSD_attack_prob function
-        # TODO: Remove the overstretched NTRU file.
-
-        # r = simulator(d=d, n=params.n, q=params.q, beta=beta, xi=xi, tau=tau)
-        # lhs = params.Xe.stddev**2 * (beta - 1)
-        # predicate = r[d - beta] > lhs
-        if not beta:
-            beta = d
-            predicate = False
+        if dsl_logvol==None:
+            dsl_logvol = PrimalDSD.DSL_logvol(params.n, params.Xs.stddev**2, ntru=params.ntru_type)
         
-        else: 
-            predicate = True
+        B_shape = [log(r_)/2 for r_ in simulator(d, params.n, params.q, beta, xi=xi, tau=tau)]
 
-        return costf(red_cost_model, beta, d, predicate=predicate)
+        dsli_vols = PrimalDSD.DSLI_vols(dsl_logvol, B_shape)
+
+        prob_all_not = RR(1.)
+        prob_pos = np.zeros(2*params.n, dtype='double')
+        for i in range(1, params.n+1):
+            s = params.n + i
+
+            dslv_len = PrimalDSD.log_gh(i, dsli_vols[s])
+            sigma_sq = exp(2*dslv_len)/s
+
+            if sigma_sq > 10**10:
+                prob_pos[s-beta] = 0.
+                continue
+
+            norm_threshold = exp(2*(B_shape[s-beta]))/sigma_sq
+            proba_one = chisquared_table[beta].cum_distribution_function(norm_threshold)
+
+            if proba_one <= 10e-8:
+                continue
+
+            # account for pulling back probability if beta small
+            if beta <= 20:
+                for j in range(2, int(s/beta+1)):
+                    if proba_one < 10**(-6):
+                        proba_one = 0.
+                        break
+                    ind = s - j*(beta-1)-1
+                    norm_bt = exp(2*B_shape[ind])/sigma_sq
+                    norm_b2 = exp(2*B_shape[ind+beta-1])/sigma_sq
+                    proba_one *= conditional_chi_squared(beta-1,s-ind-(beta-1), norm_bt, norm_b2)
+
+            prob_pos[s-beta] = proba_one
+            prob_all_not *= max(1.-proba_one, 0.)
+            Logging.log("dsd", log_level+1, f"Pr[dsd, {beta}] = {prob_all_not}")
+
+        return RR(1.-prob_all_not), prob_pos
+
 
     def __call__(
         self,
@@ -74,28 +174,89 @@ class PrimalDSD():
         **kwds,      
     ):
         """
-        TODO Add parameter description
+        Estimate cost of solving (overstretche) NTRU using the Dense sublattice.
+        Code is adapted from Léo Ducas and Wessel van Woerden.
+        See https://github.com/WvanWoerden/NTRUFatigue for original source
+
+        :param params: NTRU parameters.
+        :param red_cost_model: How to cost lattice reduction.
+        :param red_shape_model: How to model the shape of a reduced basis.
+        :return: A cost dictionary.
+
+        The returned cost dictionary has the following entries:
+
+        - ``rop``: Total number of word operations (≈ CPU cycles).
+        - ``red``: Number of word operations in lattice reduction.
+        - ``δ``: Root-Hermite factor targeted by lattice reduction.
+        - ``β``: BKZ block size.
+        - ``d``: Lattice dimension.
+
+        EXAMPLE::
+
+            >>> from estimator import *
+            >>> NTRU.primal_dsd(schemes.NTRUHPS2048509Enc)                                                                                                                                               
+            rop: ≈2^157.1, red: ≈2^157.1, δ: 1.003645, β: 453, d: 1016, tag: dsd
+
+            NOTE: Non-overstretched parameters (where the probability of Dense sublattice
+            discovery is 0) will return β = d.
+
+            >>> params = NTRU.Parameters(n=113, q=512, Xs=ND.UniformMod(3), Xe=ND.UniformMod(3))                                                                                                         
+            >>> NTRU.primal_dsd(params, red_shape_model="zgsa")                                                                                                                                          
+            rop: ≈2^41.3, red: ≈2^41.3, δ: 1.012468, β: 42, d: 226, tag: dsd
+
+            >>> NTRU.primal_dsd(params, red_shape_model="cn11")                                                                                                                                          
+            rop: ≈2^41.2, red: ≈2^41.2, δ: 1.012468, β: 41, d: 226, tag: dsd
+
+            >>> NTRU.primal_dsd(params, red_shape_model=Simulator.CN11)                                                                                                                                          
+            rop: ≈2^41.2, red: ≈2^41.2, δ: 1.012468, β: 41, d: 226, tag: dsd
+ 
+        The success condition was formulated in [EC:KF17] and further refined in [AC:DvW21]
         """
         params = NTRUParameters.normalize(params)
-
+        # allow for a larger embedding lattice dimension: Bai and Galbraith
+        m = params.m + params.n if params.Xs <= params.Xe else params.m
+ 
         try:
             red_shape_model = simulator_normalize(red_shape_model)
         except ValueError:
             pass
         
-        # Define Basis Shape Function for NTRU fatigue estimator
-        xi = PrimalUSVP._xi_factor(params.Xs, params.Xe)
-        zshapef = partial(red_shape_model,
-                    d=(params.n + params.m),
-                    xi=xi, tau=None)
-        
-        # Use NTRU fatigue Estimator for DSD event hardness
-        beta, sk_prob, dsd_prob, _ = combined_attack_prob(params.q, params.n, sk_variance=params.Xs.stddev**2, ntru=params.ntru_type, verbose=False,
-                                             only="DSD", zshapef=zshapef)
+        if params.n > max_n_cache:
+            raise ValueError("Please increase the hardcoded value of max_n_cache to run the predictor for such large n")
 
-        # print(beta, sk_prob, dsd_prob)
-        cost = self.cost_simulator(int(beta), params, red_shape_model, m=params.m, 
-                                   d=(params.m + params.n), red_cost_model=red_cost_model)
+        remaining_proba = RR(1.)
+        average_beta = RR(0.)
+        total_DSD_prob = RR(0.)
+        DSD_prob = RR(0.)
+        prob_pos_total = np.zeros(2*params.n, dtype='double')
+
+        for beta in range(2,params.n):
+            tours = floor(params.n**2 / beta**2)+3
+
+            DSD_prob, DSD_prob_pos = self.prob_dsd(beta, params, red_shape_model, m=m, red_cost_model=red_cost_model, log_level=log_level)
+
+            if DSD_prob > 10e-8:
+                for t in range(tours):
+                    for i in range(2*params.n):
+                        prob_pos = DSD_prob_pos[i]
+                        average_beta += RR(beta) * remaining_proba * prob_pos
+                        prob_pos_total[i] += remaining_proba * prob_pos
+                        total_DSD_prob += remaining_proba * prob_pos
+                        remaining_proba *= (1.-prob_pos)
+            
+                Logging.log("dsd", log_level+1, "β= %d,\t pr=%.4e, \t rem-pr=%.4e"%(beta, DSD_prob, remaining_proba))
+            if remaining_proba < 0.001:
+                average_beta += beta * remaining_proba
+                break
+
+        if not average_beta:
+            average_beta = m
+            predicate = False
+        
+        else: 
+            predicate = True
+
+        cost = costf(red_cost_model, RR(average_beta), m, predicate=predicate)
         cost["tag"] = "dsd"
         cost["problem"] = params
         return cost.sanity_check()
