+++ estimator/prob.py
@@ -2,7 +2,43 @@
 from sage.all import binomial, ZZ, log, ceil, RealField, oo, exp, pi
 from sage.all import RealDistribution, RR, sqrt, prod, erf
 from .nd import sigmaf
+from .conf import max_n_cache
 
+
+chisquared_table = {i: None for i in range(2*max_n_cache+1)}
+for i in range(2*max_n_cache+1):
+    chisquared_table[i] = RealDistribution('chisquared', i)
+
+def conditional_chi_squared(d1, d2, lt, l2):
+    """
+    Probability that a gaussian sample (var=1) of dim d1+d2 has length at most
+    lt knowing that the d2 first cordinates have length at most l2
+    """
+    D1 = chisquared_table[d1].cum_distribution_function
+    D2 = chisquared_table[d2].cum_distribution_function
+    l2 = RR(l2)
+
+    PE2 = D2(l2)
+    # In large dim, we can get underflow leading to NaN
+    # When this happens, assume lifting is successfully (underestimating security)
+    if PE2==0:
+        raise ValueError("Numerical underflow in conditional_chi_squared")
+
+    steps = 5 * (d1 + d2)
+
+    # Numerical computation of the integral
+    proba = 0.
+    for i in range(steps)[::-1]:
+        l2_min = i * l2 / steps
+        l2_mid = (i + .5) * l2 / steps
+        l2_max = (i + 1) * l2 / steps
+
+        PC2 = (D2(l2_max) - D2(l2_min)) / PE2
+        PE1 = D1(lt - l2_mid)
+
+        proba += PC2 * PE1
+
+    return proba
 
 def mitm_babai_probability(r, stddev, q, fast=False):
     """
